[#38] Dummy Data 이동 및 DummyData 모아둠
